PYbackend/ChatBot/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
sentence-transformers>=3.1.1
faiss-cpu>=1.8.0
python-dotenv>=1.0.1